# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BottleIQ API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/stores": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Stores",
        "operationId": "stores_stores_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StoreView"
                  },
                  "type": "array",
                  "title": "Response Stores Stores Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Create Store",
        "operationId": "create_store_stores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vendors": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Vendors",
        "operationId": "vendors_vendors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VendorView"
                  },
                  "type": "array",
                  "title": "Response Vendors Vendors Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Create Vendor",
        "operationId": "create_vendor_vendors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Vendor Vendors Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vendors/{vendor_id}": {
      "patch": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Update Vendor",
        "operationId": "update_vendor_vendors__vendor_id__patch",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Vendor Vendors  Vendor Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Inventory",
        "operationId": "inventory_recommendations_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Window"
            }
          },
          {
            "name": "target_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 7,
              "default": 21,
              "title": "Target Days"
            }
          },
          {
            "name": "service_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 0.999,
              "minimum": 0.8,
              "default": 0.95,
              "title": "Service Level"
            }
          },
          {
            "name": "dead_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Dead Days"
            }
          },
          {
            "name": "slow_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Slow Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Metrics"
                  },
                  "title": "Response Inventory Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Inventory",
        "operationId": "inventory_products_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Window"
            }
          },
          {
            "name": "target_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 7,
              "default": 21,
              "title": "Target Days"
            }
          },
          {
            "name": "service_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 0.999,
              "minimum": 0.8,
              "default": 0.95,
              "title": "Service Level"
            }
          },
          {
            "name": "dead_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Dead Days"
            }
          },
          {
            "name": "slow_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Slow Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Metrics"
                  },
                  "title": "Response Inventory Products Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Inventory",
        "operationId": "inventory_inventory_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Window"
            }
          },
          {
            "name": "target_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 7,
              "default": 21,
              "title": "Target Days"
            }
          },
          {
            "name": "service_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 0.999,
              "minimum": 0.8,
              "default": 0.95,
              "title": "Service Level"
            }
          },
          {
            "name": "dead_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Dead Days"
            }
          },
          {
            "name": "slow_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Slow Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Metrics"
                  },
                  "title": "Response Inventory Inventory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Product Detail",
        "operationId": "product_detail_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metrics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Update Product",
        "operationId": "update_product_products__product_id__patch",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Product Products  Product Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Alerts",
        "operationId": "alerts_alerts_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  },
                  "title": "Response Alerts Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/generate": {
      "post": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Generate",
        "operationId": "generate_recommendations_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Metrics"
                  },
                  "type": "array",
                  "title": "Response Generate Recommendations Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard": {
      "get": {
        "tags": [
          "Inventory intelligence"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_dashboard_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Alert": {
        "properties": {
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "financial_impact": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financial Impact"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "title",
          "description",
          "product_id",
          "product_name",
          "financial_impact",
          "recommended_action"
        ],
        "title": "Alert"
      },
...
      "GenerateInput": {
        "properties": {
          "window": {
            "type": "integer",
            "title": "Window",
            "default": 60
          },
          "target_days": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 7.0,
            "title": "Target Days",
            "default": 21
          },
          "service_level": {
            "type": "number",
            "maximum": 0.999,
            "minimum": 0.8,
            "title": "Service Level",
            "default": 0.95
          },
          "dead_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 30.0,
            "title": "Dead Days",
            "default": 90
          },
          "slow_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 30.0,
            "title": "Slow Days",
            "default": 90
          },
          "store_id": {
            "type": "string",
            "title": "Store Id"
          }
        },
        "type": "object",
        "required": [
          "store_id"
        ],
        "title": "GenerateInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductEdit": {
        "properties": {
          "default_vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Vendor Id"
          },
          "units_per_case": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Units Per Case"
          },
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Category"
          },
          "brand": {
            "type": "string",
            "maxLength": 120,
            "title": "Brand"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "units_per_case",
          "category",
          "brand"
        ],
        "title": "ProductEdit"
      },
...
      "StoreInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "America/Los_Angeles"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StoreInput"
      },
      "StoreView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "timezone"
        ],
        "title": "StoreView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "default_lead_time_days": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Default Lead Time Days",
            "default": 4
          },
          "minimum_order_amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Order Amount"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "VendorInput"
      },
      "VendorView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "default_lead_time_days": {
            "type": "integer",
            "title": "Default Lead Time Days"
          },
          "minimum_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Order Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "default_lead_time_days",
          "minimum_order_amount"
        ],
        "title": "VendorView"
      }
    }
  }
}